refactor : 컴파운드 패턴에서 일반 컴포넌트 방식으로 변경
