+++ apps/wow-docs/app/page.tsx
@@ -1,8 +1,7 @@
-"use client";
-
 import Checkbox from "wowds-ui/Checkbox";
 import Chip from "wowds-ui/Chip";
 import DropDown from "wowds-ui/DropDown";
+import DropDownOption from "wowds-ui/DropDownOption";
 import RadioButton from "wowds-ui/RadioButton";
 import RadioGroup from "wowds-ui/RadioGroup";
 import Switch from "wowds-ui/Switch";
@@ -18,8 +17,8 @@ const Home = () => {
         <RadioButton label="2학년" value="2학년" />
       </RadioGroup>
       <DropDown placeholder="목록을 입력하세요">
-        <DropDown.Option text="option 1" value="option 1" />
-        <DropDown.Option text="option 1" value="option 2" />
+        <DropDownOption text="option 1" value="option 1" />
+        <DropDownOption text="option 1" value="option 2" />
       </DropDown>
     </>
   );
